requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2

# Dataset downloading
datasets>=2.14.0

# Intent classification with spaCy
spacy>=3.7.0

# OpenAI API integration (optional - install if using GPT-4)
openai>=1.0.0
python-dotenv>=1.0.0

# ML-based classifiers (optional - install if using)
# transformers>=4.30.0
# torch>=2.0.0
# nemoguardrails>=0.5.0

# Statistical analysis
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
